# Make the NEWOBJ API narrower