--- a/ractor_core.h
+++ b/ractor_core.h
@@ -1,3 +1,4 @@
+#include "internal/gc.h"
 #include "ruby/ruby.h"
 #include "ruby/ractor.h"
 #include "vm_core.h"