--- a/variable.h
+++ b/variable.h
@@ -10,6 +10,8 @@
 
 /* per-object */
 
+#include "shape.h"
+
 struct gen_ivtbl {
 #if !SHAPE_IN_BASIC_FLAGS
     uint16_t shape_id;
@@ -20,7 +22,6 @@ struct gen_ivtbl {
 
 int rb_ivar_generic_ivtbl_lookup(VALUE obj, struct gen_ivtbl **);
 
-#include "shape.h"
 #if !SHAPE_IN_BASIC_FLAGS
 shape_id_t rb_generic_shape_id(VALUE obj);
 #endif