--- a/vm_core.h
+++ b/vm_core.h
@@ -667,11 +667,6 @@ typedef struct rb_vm_struct {
     VALUE mark_object_ary;
     const VALUE special_exceptions[ruby_special_error_count];
 
-    /* object shapes */
-    rb_shape_t *shape_list;
-    rb_shape_t *root_shape;
-    shape_id_t next_shape_id;
-
     /* load */
     VALUE top_self;
     VALUE load_path;