… common.mk
… ext/objspace/depend
… ext/ripper/depend
… ext/socket/depend
… yjit/src/cruby_bindings.inc.rs