Since the introduction of variable width allocation with `RVARGC` there are a lot of different `*NEWOBJ*` macros. Currently there are:

* `RB_RVARGC_NEWOBJ_OF`
* `RB_RVARGC_EC_NEWOBJ_OF`
* `RB_NEWOBJ_OF`, an alias of `RB_RVARGC_NEWOBJ_OF`
* `RB_EC_NEWOBJ_OF`, an alias of `RB_RVARGC_EC_NEWOBJ_OF`
* `NEWOBJ_OF`, an alias of `RB_RVARGC_NEWOBJ_OF`
* `RVARGC_NEWOBJ_OF`, an alias of `RB_RVARGC_NEWOBJ_OF`

This PR merges `RB_RVARGC_NEWOBJ_OF`,`RB_RVARGC_EC_NEWOBJ_OF`, `RVARGC_NEWOBJ_OF`, `RB_NEWOBJ_OF` and `NEWOBJ_OF` into a single macro that takes an execution context as an argument (which can be `NULL`; if so then the current execution context is found and used using `GET_EC()`). 

The resulting single macro has been named `NEWOBJ_OF` to reflect that it should now be the only way of creating new objects.

Both `RB_NEWOBJ_OF` and `NEWOBJ_OF` have seperate implementations that are part of the public API exposed in `include/ruby/internal/newobj.h` so they are available to extension authors. These have not been modified in any way.

Note that implementing this PR required `gc.h` to include `vm_core.h`. This was not possibly because `vm_core.h` included `shape.h`, which included `gc.h` creating a circular dependency.

To address this I have removed the `shape_list`, `next_shape_id` and `root_shape` members from `rb_vm_struct` and instead promoted them to their own type: `rb_shape_tree_t`. This is defined as a global object accessible using `rb_shape_tree_ptr` and the `GET_SHAPE_TREE()` function macro. 

In practice, this means that `GET_VM()->shape_list` etc are now replaced with `GET_SHAPE_TREE()->shape_list`.