@nobu 2023-03-01
In the commit "Remove RB_RVARGC_EC_NEWOBJ_OF and RB_EC_NEWOBJ_OF", "teh" seems a typo.