@ko1 2023-03-20
> I'm not a huge fan of passing in the EC to `RVARGC_NEWOBJ_OF` since the EC is just an implementation detail of the current GC, so in the future we should think about how we can get rid of passing the EC into the GC.

The background to pass the `ec` is here:

* `ec` is needed for Ractor systems (from 3.0)
* Getting `ec` from TLS (thread-local storage) can have overhead
* Therefore I made a plan to migrate all C-API to have the first parameter `ec` (`rbc_str_new(c, ...)` for example, in this case `c` is "context"). It is like `mrb_` APIs of mruby.
* This is why `Primitive` for builtin methods receive the `ec` parameter from VM.

However, now C-compiler specific TLS feature (`__thread` on gcc for example https://gcc.gnu.org/onlinedocs/gcc/Thread-Local.html) provides nice performance on current Ruby (as I observed).

So if you measured at least microbenchmarks the following:

* with passing `ec` (current macro)
* without passing `ec` (and `GET_EC()` everytime)

and if we don't have a difference, I agree to avoid passing the `ec`.