@ko1 2023-03-20
> (which can be NULL; if so then the current execution context is found and used using GET_EC()).

For me, this kind of branch should be avoided to reduce branches.