@eightbitraptor 2023-03-28
I got around to looking at these benchmarks recently, and to summarise the situation so far:

I refactored the NEWOBJ api's to reduce complexity at the call sites. This has resulted in an api that requires us to either pass an `ec` or `NULL` to the `NEWOBJ_OF` internal macro.

@peterzhu2118 had a concern that requiring call sites to pass the `ec` was leaking GC abstractions.

@ko1 was concerned that _not_ passing the `ec` and instead relying on `GET_EC` might be a performance problem because of the overhead incurred by accessing a thread local every time.

Additionally @ko1 was concerned that the extra branch point required now in the `NEWOBJ_OF` macro might also cause a performance impact.

I built a micro-benchmark, designed to produce many `putstring` instructions ran it across several variations of this feature.

This is the benchmark:

```
eval("def string_literal; foo = [#{ 10_000.times.map { "'test'" }.join(",") }]; nil; end")

before = Process.clock_gettime(Process::CLOCK_MONOTONIC)
10000.times { string_literal }
p Process.clock_gettime(Process::CLOCK_MONOTONIC) - before
```

These are the branches I used, and how they compare:

* `master` - passes EC explicitly in the cases of the `putstring` and `newarray` instructions. So that we don't have to use `GET_EC` to retreive the thread local ec variable.

* `mvh-narrow-newobj-api` - Based on `master` but with the newobj paths in `gc.c` and the `NEWOBJ_OF` macros refactored to reduce paths through the code. `ec` still must be explicitly passed, although this is now decided using a branch generated by the macro, rather than by using completely seperate code paths as in `master`.

* `mvh-narrow-newobj-api-macro-branch` - Same as `mvh-narrow-newobj-api`. but additionally removes the branch point, by using some pre-processor string concatenation and dynamic macro name generation to handle the branch at the pre-processor level.

* `mvh-narrow-newobj-api-remove-ec` - This builds on top of the refactors carried out in `mvh-narrow-newobj-api` but also completely removes the code path that passes the `ec` in the `putstring` and `newarray` instructions, relying instead on always getting the `ec` from the thread. This eliminates all the extra code paths required to check the presence of the `ec`.

## results

<img width="1337" alt="Screenshot 2023-03-28 at 17 21 29" src="https://user-images.githubusercontent.com/31869/228316599-9c000760-d84d-4827-aa8e-8086017174f5.png">

|     | master     | mvh-narrow-newobj-api | mvh-narrow-newobj-api-remove-ec | mvh-narrow-newbj-api-macro-branch |
| --- | ---------- | --------------------- | ------------------------------- | --------------------------------- |
| 1   | 3.50582845 | 3.51671825            | 3.42554660                      | 3.62113722                        |
| 2   | 3.52347104 | 3.50921376            | 3.42405504                      | 3.63043113                        |
| 3   | 3.50835028 | 3.51863772            | 3.48882321                      | 3.62117951                        |
| 4   | 3.50525342 | 3.51858662            | 3.43421950                      | 3.60670734                        |
| 5   | 3.50967737 | 3.51790606            | 3.42857820                      | 3.60528875                        |
| avg | 3.51051611 | 3.51621248            | 3.44024451                      | 3.61694879                        |

Benchmarking was done on Fedora 37 on a Ryzen 3600, running at a fixed 3.6GHz, with each process restricted to a single CPU core and with ASLR disabled, to try and reduce variance.

We can see from these results that `mvh-narrow-newobj-api-remove-ec` is consitently faster than all the other branches. From this I conclude that the compiler optimisations for thread locals are good enough now to offset the performance hit caused by repeatedly accessing the thread local `ec` using `GET_EC`.

So I would like to merge the patches that remove the `ec` from the internal `NEWOBJ_OF` into this branch and merge it.