@eightbitraptor 2023-03-28
One thing I don't have a good answer for at this time is why `mvh-narrow-newobj-api-macro-branch` is slower than `mvh-narrow-newobj-api`, given that it's the same functionality but missing a branch.

We can even see this by looking at the generated code for the `rb_ec_str_ressurrect` function with `objdump`:

```
objdump -M intel --disassemble=rb_ec_str_resurrect --no-addresses miniruby | wc -l
```

* on the `mvh-narrow-newobj-api` the generated asm is 193 lines
* on the `mvh-narrow-newobj-api-macro-branch` branch the generated asm is 181 lines

In addition, in `mvh-narrow-newobj-api` we can clearly see the calls to `rb_current_ec` where it gets the thread local

```
❯ objdump -M intel --disassemble=rb_ec_str_resurrect --no-addresses miniruby | grep ruby_current_ec
        48 8b 05 61 5f 2b 00    mov    rax,QWORD PTR [rip+0x2b5f61]        # <ruby_current_ec@@Base+0x50dca0>
        48 8b 05 29 5f 2b 00    mov    rax,QWORD PTR [rip+0x2b5f29]        # <ruby_current_ec@@Base+0x50dca0>
```

So given that it generates more code, and that extra code is fetching the `ec`, I can't explain why it's faster.

I don't think this is anything to worry about, given that removing the `ec` is faster than both of these branches. But it's a curiosity I can't explain.