@tenderlove 2023-03-29
> We can see from these results that `mvh-narrow-newobj-api-remove-ec` is consitently faster than all the other branches. From this I conclude that the compiler optimisations for thread locals are good enough now to offset the performance hit caused by repeatedly accessing the thread local `ec` using `GET_EC`.

We should probably try this benchmark with more than 1 thread live in the system.  I'd guess we'll see stalls when there are multiple threads?